+++ packages/sit-onyx/src/components/OnyxButton/OnyxButton.ct.tsx
@@ -60,6 +60,14 @@ test("should truncate text", async ({ mount }) => {
   await expect(component).toHaveScreenshot("truncation-ellipsis.png");
 });
 
+test("should render skeleton", async ({ mount }) => {
+  // ARRANGE
+  const component = await mount(<OnyxButton label="Test label" skeleton />);
+
+  // ASSERT
+  await expect(component).toHaveScreenshot("skeleton.png");
+});
+
 const STATES = {
   state: ["default", "disabled", "icon"],
   variation: ["primary", "secondary", "danger"],