+++ packages/sit-onyx/src/components/OnyxRadioButton/OnyxRadioButton.stories.ts
@@ -40,3 +40,13 @@ export const WithTruncation = {
   },
   decorators: [createTruncationDecorator("12rem")],
 } satisfies Story;
+
+/**
+ * This example shows a skeleton radio button.
+ */
+export const Skeleton = {
+  args: {
+    ...Default.args,
+    skeleton: true,
+  },
+} satisfies Story;