+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sit-onyx
 
+## 1.0.0-alpha.32
+
+### Minor Changes
+
+- c147e3d: feat: add property `skeleton` to OnyxButton, OnyxCheckbox, OnyxRadioButton and OnyxSwitch
+
+## 1.0.0-alpha.31
+
+### Minor Changes
+
+- 69f1569: feat: add truncation for OnyxButton, OnyxCheckbox, OnyxRadioButton and OnyxSwitch
+
 ## 1.0.0-alpha.30
 
 ### Patch Changes